refactor(admin): extract AdminDebuggingSection into 4 focused sub-panel components
Pure structural refactor — no behaviour changes. The 949-line monolithic
file is split into 4 self-contained panel files plus a thin composition
shell (~30 lines). All 13 characterisation tests and 5 denial panel tests
pass with zero assertion changes.

Extracted files:
- AdminErrorTimelinePanel.jsx  (ErrorLogCentrePanel + OccurrenceTimeline + ErrorEventDetailsDrawer)
- AdminRequestDenialsPanel.jsx (DenialLogPanel + DENIAL_REASON_OPTIONS + DENIAL_REASON_LABEL_MAP)
- AdminDebugBundlePanel.jsx    (DebugBundlePanel + DebugBundleSectionTable + DebugBundleResult)
- AdminLearnerSupportPanel.jsx (LearnerSupportPanel)
